pyproject.toml
[tool.poetry]
name = "skybank-project"
version = "0.1.0"
description = "cousework for skypro"
authors = ["Anton <fynjy-666@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
openpyxl = "^3.1.5"
pytest = "^8.3.2"


[tool.poetry.group.lint.dependencies]
black = "^24.4.2"
mypy = "^1.10.1"
flake8 = "^7.1.0"
isort = "^5.13.2"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^5.0.0"
pytest = "^8.3.2"
freezegun = "^1.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''

[tool.isort]
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
ignore_missing_imports = true
exclude = 'venv'
